[types] Update directory structure for account_config
