+++ types/src/account_config/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod constants;
+pub mod events;
+pub mod resources;
+
+pub use constants::*;
+pub use events::*;
+pub use resources::*;
