+++ types/src/account_config/resources/balance.rs
@@ -1,0 +1,56 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::{
+    access_path::{AccessPath, Accesses},
+    account_config::constants::{lbr_type_tag, ACCOUNT_MODULE_NAME, CORE_CODE_ADDRESS},
+    language_storage::{StructTag, TypeTag},
+    move_resource::MoveResource,
+};
+#[cfg(any(test, feature = "fuzzing"))]
+use proptest_derive::Arbitrary;
+use serde::{Deserialize, Serialize};
+
+/// The balance resource held under an account.
+#[derive(Debug, Serialize, Deserialize)]
+#[cfg_attr(any(test, feature = "fuzzing"), derive(Arbitrary))]
+pub struct BalanceResource {
+    coin: u64,
+}
+
+impl BalanceResource {
+    pub fn new(coin: u64) -> Self {
+        Self { coin }
+    }
+
+    pub fn coin(&self) -> u64 {
+        self.coin
+    }
+
+    // TODO/XXX: remove this once the MoveResource trait allows type arguments to `struct_tag`.
+    pub fn struct_tag_for_currency(currency_typetag: TypeTag) -> StructTag {
+        StructTag {
+            address: CORE_CODE_ADDRESS,
+            name: BalanceResource::struct_identifier(),
+            module: BalanceResource::module_identifier(),
+            type_params: vec![currency_typetag],
+        }
+    }
+
+    // TODO: remove this once the MoveResource trait allows type arguments to `resource_path`.
+    pub fn access_path_for(currency_typetag: TypeTag) -> Vec<u8> {
+        AccessPath::resource_access_vec(
+            &BalanceResource::struct_tag_for_currency(currency_typetag),
+            &Accesses::empty(),
+        )
+    }
+}
+
+impl MoveResource for BalanceResource {
+    const MODULE_NAME: &'static str = ACCOUNT_MODULE_NAME;
+    const STRUCT_NAME: &'static str = "Balance";
+
+    fn type_params() -> Vec<TypeTag> {
+        vec![lbr_type_tag()]
+    }
+}
