+++ types/src/account_config/resources/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod account;
+pub mod association_capability;
+pub mod balance;
+pub mod currency_info;
+
+pub use account::*;
+pub use association_capability::*;
+pub use balance::*;
+pub use currency_info::*;
